+++ lib/model/player.dart
@@ -1,3 +1,4 @@
+import 'package:fs_score_card/model/french_driving_round_attributes.dart';
 import 'package:fs_score_card/model/phases.dart';
 import 'package:fs_score_card/model/round_states.dart';
 import 'package:fs_score_card/model/scores.dart';
@@ -7,12 +8,17 @@ class Player {
   Player({required this.name, required int maxRounds, required int numPhases})
     : scores = Scores(maxRounds),
       phases = Phases(numPhases),
+      frenchDrivingAttributes = List.generate(
+        maxRounds,
+        (_) => FrenchDrivingRoundAttributes(),
+      ),
       roundStates = RoundStates(maxRounds);
 
   Player.withData({
     required this.name,
     required this.scores,
     required this.phases,
+    required this.frenchDrivingAttributes,
     RoundStates? roundStates,
   }) : roundStates = roundStates ?? RoundStates(scores.roundScores.length);
 
@@ -21,11 +27,21 @@ class Player {
     : name = json['name'] as String,
       scores = Scores.fromJson(json['scores'] as List<dynamic>),
       phases = Phases.fromJson(json['phases'] as List<dynamic>),
+      frenchDrivingAttributes =
+          (json['frenchDrivingAttributes'] as List<dynamic>?)
+              ?.map(
+                (e) => FrenchDrivingRoundAttributes.fromJson(
+                  e as Map<String, dynamic>,
+                ),
+              )
+              .toList() ??
+          [],
       roundStates = RoundStates.fromJson(json['roundStates'] as List<dynamic>);
 
   final String name;
   final Scores scores;
   final Phases phases;
+  final List<FrenchDrivingRoundAttributes> frenchDrivingAttributes;
   final RoundStates roundStates;
 
   int get totalScore => scores.total;
@@ -35,6 +51,9 @@ class Player {
     'name': name,
     'scores': scores.toJson(),
     'phases': phases.toJson(),
+    'frenchDrivingAttributes': frenchDrivingAttributes
+        .map((e) => e.toJson())
+        .toList(),
     'roundStates': roundStates.toJson(),
     'totalScore': totalScore,
   };